Don't format phone numbers.
Numbers should not be altered in any way. Otherwise, it would defeat the purpose of this gem. For example, if I want to test that an invalid number raises an error, and if this gem alters that number so that it becomes valid, I will never be able to test that scenario.